Added sharedStructs.go & finished the remaining .go's and _types.go's.
