+++ cart.go
@@ -4,8 +4,8 @@ const (
 	createOrFetchCartEndpoint = "/checkout/cart"
 	deleteCartEndpoint        = "/checkout/cart"
 	addItemCartEndpoint       = "/checkout/cart/line-item"
-	updateItemCartEndpoint    = "/checkout/cart/line-item"
 	removeItemCartEndpoint    = "/checkout/cart/line-item"
+	updateItemCartEndpoint    = "/checkout/cart/line-item"
 )
 
 // CartClient handles all requests regarding the cart.
@@ -13,7 +13,7 @@ type CartClient struct {
 	client *Client
 }
 
-//CreateOrFetch creates or fetches a new cart.
+// CreateOrFetch creates or fetches a new cart.
 func (client *CartClient) CreateOrFetch() (*Cart, error) {
 	result := new(Cart)
 
@@ -24,7 +24,7 @@ func (client *CartClient) CreateOrFetch() (*Cart, error) {
 	return result, nil
 }
 
-//Delete deletes a cart.
+// Delete deletes a cart.
 func (client *CartClient) Delete() (*DeleteCartResult, error) {
 	result := new(DeleteCartResult)
 
@@ -35,7 +35,7 @@ func (client *CartClient) Delete() (*DeleteCartResult, error) {
 	return result, nil
 }
 
-//AddItems adds new items into the cart.
+// AddItems adds new items into the cart.
 func (client *CartClient) AddItems(add AddItemCart) (*Cart, error) {
 	result := new(Cart)
 
@@ -46,7 +46,18 @@ func (client *CartClient) AddItems(add AddItemCart) (*Cart, error) {
 	return result, nil
 }
 
-//UpdateItems updates items inside the cart.
+// RemoveItems removes items from the cart and recalculates it.
+func (client *CartClient) RemoveItems() (*Cart, error) {
+	result := new(Cart)
+
+	if err := client.client.performDelete(removeItemCartEndpoint, result); err != nil {
+		return nil, err
+	}
+
+	return result, nil
+}
+
+// UpdateItems updates items inside the cart.
 func (client *CartClient) UpdateItems(update UpdateItemCart) (*Cart, error) {
 	result := new(Cart)
